requirements.txt
torch
torchvision
pillow
zstd